refactor(console): move getSubscription into ApplicationSubscriptionService
